Bug fixes
Fixed a bug when the projectile refused to launch up, no matter what vector i've set. The issue was that the projectile was spawning in player collision and it had collision with char.

Made it so the projectile is launched on button release.

Mad it so the longer you hold the throw - the more powerful it is. (might wanna add visuals like in worms)